Assignment3
Go to our landing page in the course project and add a “margin-top” to our image (“#product-overview”) to ensure, that both the image and the slogan („Get the freedom you deserve“) are no longer hidden below the navigation bar.
Find a way to position the slogan in the left lower corner of the image using “px”. The slogan should have some space to the left and to the bottom but do NOT add a margin to it.
Change the positioning context of the slogan (challenge: What was the positioning context so far? What is the new positioning context?). Remember how we created the badge on the “Packages” page…
Adjust the position of the slogan using “%” values.
Switch to the “Packages” page.
Add a “z-index” of 10 to the badge (“.package__badge”). Scroll down on the webpage, what happens to the badge and what causes this issue?
Fix the issue encountered using the “z-index”. Make sure that the highest “z-index” on the entire page is the “z-index” of the badge (hint: Maybe the Stacking Context is a helpful concept in this situation, so you might have to add the “position” property again).